+++ client/styles/style.css
@@ -1,24 +1,25 @@
 body{
-    padding: 50px;
-    font-size: 14px;
-    font-family: "Lucida Grande", Helvetica, Arial, sans-serif;
-  }
-  
-  a{
-    color: #00B7FF;
-  }
-  
-  h1{
-    color: teal;
-  }
+  padding: 50px;
+  font-size: 14px;
+  font-family: "Lucida Grande", Helvetica, Arial, sans-serif;
+}
 
-  #brand-logo-img {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-  }
-  
-  #brand-logo-img img {
-    margin-right: 10px; /* Add some space between the image and text if desired */
-  }
+a{
+  color: #00B7FF;
+}
+
+h1 {
+  color: teal;
+}
+
+.brand-logo-img {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+}
+
+.brand-logo-img img {
+  margin-right: 10px; /* Add some space between the image and text if desired */
+  border-radius: 25px;
+}
