+++ server/domains/about/about.controller.js
@@ -1,0 +1,11 @@
+// Actions methods
+
+// GET '/user/about'
+const about = (req, res) => {
+    res.send("🚧 UNDER CONSTRUCTION '/user/about' 🚧");
+};
+
+// Controlador Home
+export default {
+  about,
+};
