+++ server/domains/project/project.controller.js
@@ -1,0 +1,16 @@
+// Actions methods
+
+// GET '/user/project/["projects", "dashboard"]'
+const projects = (req, res) => {
+  res.send("🚧 UNDER CONSTRUCTION '/user/project/[projects o dashboard]' 🚧");
+};
+
+const forms = (req, res) => {
+  res.send("🚧 UNDER CONSTRUCTION '/user/project/[add-form o add]' 🚧");
+};
+
+// Controlador Home
+export default {
+  projects,
+  forms,
+};
