+++ server/domains/user/user.controller.js
@@ -1,21 +1,23 @@
-// Importando el Router de Express
-import { Router } from 'express';
+// Actions methods
 
-// Importando el controlador
-import userController from './user.controller';
-
-// Creando una isntancia del enrutador
-const router = new Router();
-
-// Enrutamos
 // GET '/user/login'
-router.get(['/login'], userController.login);
+const login = (req, res) => {
+  res.send("🚧 UNDER CONSTRUCTION '/user/login' 🚧");
+};
 
 // GET '/user/logout'
-router.get(['/logout'], userController.logout);
+const logout = (req, res) => {
+  res.send("🚧 UNDER CONSTRUCTION '/user/logout' 🚧");
+};
 
 // GET '/user/register'
-router.get(['/register'], userController.register);
+const register = (req, res) => {
+  res.send("🚧 UNDER CONSTRUCTION '/user/register' 🚧");
+};
 
-// Exporto este tramo de ruta
-export default router;
+// Controlador Home
+export default {
+  login,
+  logout,
+  register,
+};
