+++ server/router.js
@@ -1,21 +1,24 @@
+// Helps to handle http errors
 import createError from 'http-errors';
 // Importing winston logger
 import log from './config/winston';
-
 // Importando enrutador home
 import homeRouter from './domains/home/home.router';
-import userRouter from './user/user.router';
+import userRouter from './domains/user/user.router';
+import projectRouter from './domains/project/project.router';
+import aboutRouter from './domains/about/about.router';
 
 // Función que agrega rutas
 const addRoutes = (app) => {
   // Agregando enrutado de Home
   app.use('/', homeRouter);
   app.use('/user', userRouter);
-
+  app.use('/project', projectRouter);
+  app.use('/about', aboutRouter);
   // ERRORES
   // catch 404 and forward to error handler
   app.use((req, res, next) => {
-    log.info(`404 Pagina no encontrada ${req.method} ${req.originalUrl}`);
+    log.info(`404 Página no encontrada ${req.method} ${req.originalUrl}`);
     next(createError(404));
   });
 
@@ -34,4 +37,5 @@ const addRoutes = (app) => {
   return app;
 };
 
+// Exportando objeto
 export default { addRoutes };
