+++ server/views/partials/_navbar.hbs
@@ -1,20 +1,20 @@
 <!-- Dropdow para menu "Projectos" -->
 <ul id="dropdownMenu" class="dropdown-content">
   <li>
-    <a href="/projects">
+    <a href="/project">
       <i class="large material-icons" style="margin: 2px;">
         list
       </i>
-      Listar Proyectos
+      Listar
     </a>
   </li>
   <li class="divider"></li>
   <li>
-    <a href="/projects/add">
+    <a href="/project/add">
       <i class="large material-icons" style="margin: 2px;">
         add_box
       </i>
-      Crear Proyecto
+      Crear
     </a>
   </li>
 </ul>
@@ -76,7 +76,7 @@
         <!-- Dropdown Trigger -->
         <li>
           <a class="dropdown-trigger" href="#!" data-target="dropdownMenu">
-            Acciones
+            Proyectos
             <i class="material-icons right">
               arrow_drop_down
             </i>
@@ -109,7 +109,7 @@
         </li>
         <li class="divider"></li>
         <li>
-          <a href="/projects">
+          <a href="/project">
             <i class="large material-icons left">
               list
             </i>
@@ -117,7 +117,7 @@
           </a>
         </li>
         <li>
-          <a href="/projects/add">
+          <a href="/project/add">
             <i class="large material-icons left">
               add_box
             </i>
